moved yearEstablied to seperate function in helperFunctions.js
